requirements.txt
# Core dependencies
playwright>=1.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
